backend/requirements.txt
asgiref==3.11.1
bcrypt==5.0.0
Django==6.0.2
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dnspython==2.8.0
mongoengine==0.29.1
PyJWT==2.11.0
pymongo==4.16.0
sqlparse==0.5.5
mongomock==4.3.0
boto3==1.36.23
django-extensions==3.2.3
